# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "opendataAPI",
        "description": "Open data API from open dataset",
        "contact": {
            "email": "mellyniaramadhan762@gmail.com"
        },
        "license": {
            "name": "mellyniaramadhan",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/BPS/InflationRates": {
            "get": {
                "tags": [
                    "Inflation Rates"
                ],
                "summary": "Get a list of inflation rates",
                "operationId": "596b394638e959c12042a6826fecd452",
                "responses": {
                    "200": {
                        "description": "List of inflation rates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InflationRates"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inflation-rates/{id}": {
            "get": {
                "tags": [
                    "Inflation Rates"
                ],
                "summary": "Get details of an inflation rate",
                "operationId": "097c0a9cd5c9c7c380ed1837f3045d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inflation rate to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Meulaboh"
                                        },
                                        "2017PercentAmount": {
                                            "type": "string",
                                            "example": "4.76"
                                        },
                                        "2018PercentAmount": {
                                            "type": "string",
                                            "example": "0.96"
                                        },
                                        "2019PercentAmount": {
                                            "type": "string",
                                            "example": "4.28"
                                        },
                                        "2020PercentAmount": {
                                            "type": "string",
                                            "example": "4.24"
                                        },
                                        "2021PercentAmount": {
                                            "type": "string",
                                            "example": "2.07"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-15 13:47:57"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-15 13:47:57"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not Found. Inflation rate with the given ID not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inflation Rates"
                ],
                "summary": "Update an existing inflation rate",
                "operationId": "7237b6a7ef25e53b62c6ba8e06ce40f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inflation rate to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data to update an existing inflation rate",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "number",
                                        "example": 2.5
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inflation rate updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "city": {
                                            "type": "number"
                                        },
                                        "year": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found. Inflation rate with the given ID not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inflation Rates"
                ],
                "summary": "Delete an inflation rate",
                "operationId": "dee404681797a2a9d624f99868799665",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inflation rate to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inflation rate deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not Found. Inflation rate with the given ID not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inflation-rates": {
            "post": {
                "tags": [
                    "Inflation Rates"
                ],
                "summary": "Create a new inflation rate",
                "operationId": "a51d1803523e991a2f2046abcdd37852",
                "requestBody": {
                    "description": "Data to create a new inflation rate",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "number"
                                    },
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "city": "example city",
                                "year": "example year"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inflation rate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Meulaboh"
                                        },
                                        "2017PercentAmount": {
                                            "type": "string",
                                            "example": "4.76"
                                        },
                                        "2018PercentAmount": {
                                            "type": "string",
                                            "example": "0.96"
                                        },
                                        "2019PercentAmount": {
                                            "type": "string",
                                            "example": "4.28"
                                        },
                                        "2020PercentAmount": {
                                            "type": "string",
                                            "example": "4.24"
                                        },
                                        "2021PercentAmount": {
                                            "type": "string",
                                            "example": "2.07"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-15 13:47:57"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-15 13:47:57"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InflationRates": {
                "required": [
                    "City",
                    "2017PercentAmount",
                    "2018PercentAmount",
                    "2019PercentAmount",
                    "2020PercentAmount",
                    "2021PercentAmount"
                ],
                "properties": {
                    "City": {
                        "type": "string"
                    },
                    "2017PercentAmount": {
                        "type": "string"
                    },
                    "2018PercentAmount": {
                        "type": "string"
                    },
                    "2019PercentAmount": {
                        "type": "string"
                    },
                    "2020PercentAmount": {
                        "type": "string"
                    },
                    "2021PercentAmount": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}